Refactor Portfolio component and introduce new sections
- Integrated new sections: About, Skills, Projects, Experience, Testimonials, and Contact into the Portfolio component.
- Replaced the Hero section with a more dynamic HeroSection component.
- Added a ContactForm component for user inquiries.
- Enhanced UI with new background effects and improved responsiveness.
- Updated lazy-loaded data structure for better content management.
- Removed deprecated code and optimized component structure for maintainability.
