+++ constants/lazy-data.ts
@@ -22,42 +22,38 @@ export const ABOUT_CONTENT = {
 
 // Skills organized by category (pre-computed for performance)
 const RAW_SKILLS = [
-  { name: "JavaScript", category: "frontend", icon: "logos:javascript" },
-  { name: "React", category: "frontend", icon: "logos:react" },
-  { name: "Redux", category: "frontend", icon: "logos:redux" },
-  { name: "Vue", category: "frontend", icon: "logos:vue" },
-  { name: "CSS3", category: "frontend", icon: "logos:css-3" },
-  { name: "vitest", category: "testing", icon: "logos:vitest" },
-  { name: "TypeScript", category: "frontend", icon: "logos:typescript-icon" },
-  { name: "Vite", category: "frontend", icon: "logos:vitejs" },
-  { name: "TailwindCSS", category: "frontend", icon: "logos:tailwindcss-icon" },
-  { name: "AntD", category: "frontend", icon: "logos:ant-design" },
-  { name: "Ruby", category: "backend", icon: "logos:ruby" },
-  { name: "Ruby on Rails", category: "backend", icon: "logos:rails" },
-  { name: "Strapi", category: "backend", icon: "logos:strapi-icon" },
-  { name: "MongoDB", category: "backend", icon: "logos:mongodb-icon" },
-  { name: "MySQL", category: "backend", icon: "logos:mysql-icon" },
-  { name: "PostgreSQL", category: "backend", icon: "logos:postgresql" },
-  { name: "NodeJS", category: "backend", icon: "logos:nodejs-icon" },
-  { name: "Express", category: "backend", icon: "logos:express" },
-  { name: "GraphQL", category: "backend", icon: "logos:graphql" },
-  { name: "Machine Learning", category: "ai", icon: "mdi:robot" },
-  { name: "LLMs", category: "ai", icon: "mdi:brain" },
-  { name: "NLP", category: "ai", icon: "mdi:message-text" },
-  { name: "Vite", category: "testing", icon: "logos:vitejs" },
-  { name: "Jest", category: "testing", icon: "logos:jest" },
-  { name: "TDD", category: "testing", icon: "mdi:check-decagram" },
-  { name: "Playwright", category: "testing", icon: "logos:playwright" },
-  {
-    name: "Chrome Dev Tools",
-    category: "testing",
-    icon: "logos:google-chrome",
-  },
-  { name: "Git", category: "tools", icon: "logos:git-icon" },
-  { name: "GitHub", category: "tools", icon: "logos:github-icon" },
-  { name: "VS Code", category: "tools", icon: "logos:visual-studio-code" },
-  { name: "Cursor", category: "tools", icon: "ph:cursor-duotone" },
-  { name: "AWS", category: "tools", icon: "logos:aws" },
+  { name: "JavaScript", category: "frontend", icon: "☕️" },
+  { name: "React", category: "frontend", icon: "⚛️" },
+  { name: "Redux", category: "frontend", icon: "✨" },
+  { name: "Vue", category: "frontend", icon: "🟩" },
+  { name: "CSS3", category: "frontend", icon: "🎨" },
+  { name: "vitest", category: "testing", icon: "⚡" },
+  { name: "TypeScript", category: "frontend", icon: "📘" },
+  { name: "Vite", category: "frontend", icon: "📦" },
+  { name: "TailwindCSS", category: "frontend", icon: "💨" },
+  { name: "AntD", category: "frontend", icon: "🐜" },
+  { name: "Ruby", category: "backend", icon: "💎" },
+  { name: "Ruby on Rails", category: "backend", icon: "🛤️" },
+  { name: "Strapi", category: "backend", icon: "🚀" },
+  { name: "MongoDB", category: "backend", icon: "🍃" },
+  { name: "MySQL", category: "backend", icon: "🐬" },
+  { name: "PostgreSQL", category: "backend", icon: "🐘" },
+  { name: "NodeJS", category: "backend", icon: "🟢" },
+  { name: "Express", category: "backend", icon: "⚡️" },
+  { name: "GraphQL", category: "backend", icon: "📊" },
+  { name: "Machine Learning", category: "ai", icon: "🤖" },
+  { name: "LLMs", category: "ai", icon: "🧠" },
+  { name: "NLP", category: "ai", icon: "🗣️" },
+  { name: "Vite", category: "testing", icon: "⚡" },
+  { name: "Jest", category: "testing", icon: "🃏" },
+  { name: "TDD", category: "testing", icon: "✅" },
+  { name: "Playwright", category: "testing", icon: "🎭" },
+  { name: "Chrome Dev Tools", category: "testing", icon: "⚙️" },
+  { name: "Git", category: "tools", icon: "🌱" },
+  { name: "GitHub", category: "tools", icon: "🐙" },
+  { name: "VS Code", category: "tools", icon: "💻" },
+  { name: "Cursor", category: "tools", icon: "🗂️" },
+  { name: "AWS", category: "tools", icon: "☁️" },
 ];
 
 export const SKILLS_CATEGORIES = [
@@ -192,8 +188,8 @@ export const PROJECT_CONTENT = {
 export const TESTIMONIALS_DATA = [
   {
     name: "Oscar Fernández Muñoz",
-    role: "Full Stack Developer",
-    company: "Microverse",
+    role: "Software Engineer",
+    company: "Torre",
     image: "/placeholder.svg?height=80&width=80",
     content:
       "I had the pleasure of pair-programming with Miliyon during Microverse. He's detail-oriented, writes clean, accessible code, and consistently looks for UI improvements. On top of that, he's approachable and easy to collaborate with. Highly recommended!",
@@ -224,6 +220,15 @@ export const TESTIMONIALS_DATA = [
     image: "/placeholder.svg?height=80&width=80",
     content:
       "I collaborated with Miliyon on a React project at Microverse. He has solid JavaScript skills and confidently contributes valuable ideas.",
+    rating: 5,
+  },
+  {
+    name: "Felipe Oyarzo Contreras",
+    role: "Junior Software Engineer",
+    company: "Microverse",
+    image: "/placeholder.svg?height=80&width=80",
+    content:
+      "I worked with Miliyon on a Microverse capstone project, and it was a great experience. He’s a skilled problem solver and would be a valuable asset to any team.",
     rating: 5,
   },
 ];
